requirements.txt
altgraph==0.17.4
beautifulsoup4==4.12.3
greenlet==3.1.1
macholib==1.16.3
packaging==24.1
pillow==11.0.0
playwright==1.48.0
pyee==12.0.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
soupsieve==2.6
typing_extensions==4.12.2
